requirements.txt
vibecoder-secure-mcp
fastapi
uvicorn
pyyaml
cryptography
mistune
openai
langchain
redis
gitpython
python-gitlab